+++ beaker_kernel/lib/code_analysis/__init__.py
@@ -1,0 +1,17 @@
+from .analyzer import CodeAnalyzer
+from .analysis_types import AnalysisAnnotation, AnalysisCategory, AnalysisItem
+from .rules import TrustRule, ASTRule, LLMRule
+from .analysis_agent import CodeAnalysisAgent, AnalysisResult
+
+
+__all__ = [
+    "CodeAnalyzer",
+    "AnalysisAnnotation",
+    "AnalysisCategory",
+    "AnalysisItem",
+    "TrustRule",
+    "ASTRule",
+    "LLMRule",
+    "CodeAnalysisAgent",
+    "AnalysisResult",
+]
