+++ beaker_kernel/lib/code_analysis/analysis_agent.py
@@ -7,10 +7,13 @@ from pydantic import BaseModel, Field
 
 from archytas.react import ReActAgent, tool, LoopController
 from archytas.tool_utils import LoopControllerRef
-from .linter import TrustAnnotationType
+from .analyzer import AnalysisCategory
 
 
-class AnalysisObject(BaseModel):
+class AnalysisResult(BaseModel):
+    """
+    Representation of a lint-style analysis that will be generated by an LLM agent and passed to the code_analysis tool.
+    """
     code_start_line: int = Field(description="Starting line number for which the analysis applies.")
     code_end_line: int = Field(description=(
         "Ending line number for which the analysis applies. If the analysis is for a single line, this value should be "
@@ -25,10 +28,13 @@ class AnalysisObject(BaseModel):
         "Optional extra information that may be useful to the user to understand the analysis, particularly in cases "
         "where the issue or rule is vague or potentially confusing."
     ))
+    cell_id: Optional[str] = Field(description=(
+        "The ID of the cell the code is generated from, if available from the context."
+    ))
 
 
-class TrustAgent(ReActAgent):
-    """\
+class CodeAnalysisAgent(ReActAgent):
+    """
     You are a code evaluation agent, whose job it is to evaluate code that is generated by a LLM and to provide helpful
     information so that they can more easily identify potential issues with the code and to understand the code and what
     it is doing more efficiently.
@@ -54,16 +60,18 @@ class TrustAgent(ReActAgent):
             **kwargs
         )
 
-    def set_annotation_types(self, annotation_types: list[TrustAnnotationType]):
-        content = [f"""
+    def set_annotation_types(self, annotation_types: list[AnalysisCategory]):
+        content = [
+            f"""\
 Below are the types of annotations/issues that you should use when analyzing code and generating analyses.
 All generated analyses should match exactly one of these rules.
-"""]
+"""
+        ]
         for annotation_type in annotation_types:
             type_str = []
             type_str.append(f"""
 # {annotation_type.display_label} ({annotation_type.id})""".strip())
-            for annotation in annotation_type.message_types:
+            for annotation in annotation_type.analysis_items:
                 f"""\
 ## {annotation.title} ({annotation.id})
     id: {annotation.id}
@@ -73,12 +81,13 @@ All generated analyses should match exactly one of these rules.
 {annotation.description}
 ```"""
             content.append("\n\n".join(type_str))
-
-        print(content)
         self.add_context("\n\n".join(content))
 
+    async def execute(self, additional_messages = None):
+        return await super().execute(additional_messages)
+
     @tool
-    def code_analysis(self, analysis_list: list[AnalysisObject], loop: LoopControllerRef) -> list[AnalysisObject]:
+    def code_analysis(self, analysis_list: list[AnalysisResult], loop: LoopControllerRef) -> list[AnalysisResult]:
         """
         Given the code and rules above, please inspect the code for matching issues. For each issue identified, please
         generate an Analysis object
@@ -89,11 +98,11 @@ All generated analyses should match exactly one of these rules.
         If successful, this tool will immediately stop the react loop and return the analyses to the invoking process.
 
         Arguments:
-            analysis_list: A list of analysis objects indicating the type and location of the issue.
+            analysis_list (list[AnalysisObject]): A list of analysis objects indicating the type and location of the issue.
         Returns:
             list[AnalysisObject]: List of analyses provided to the tool directly to the calling function.
-
         """
+
         loop.set_state(loop.STOP_SUCCESS)
         output =  dill.dumps(analysis_list)
         # Return must be a string, and dill.dumps returns bytes so decode appropriately
